+++ Source/Utility/Mat3f.h
@@ -126,12 +126,12 @@ namespace TrenchBroom {
                 return *this;
             }
             
-            inline float& operator[] (const unsigned int index) {
+            inline float& operator[] (const size_t index) {
                 assert(index >= 0 && index < 9);
                 return v[index];
             }
             
-            inline const float& operator[] (const unsigned int index) const {
+            inline const float& operator[] (const size_t index) const {
                 assert(index >= 0 && index < 9);
                 return v[index];
             }
