+++ Source/Utility/Vec3f.h
@@ -204,14 +204,14 @@ namespace TrenchBroom {
                 return *this;
             }
             
-            inline float& operator[] (const unsigned int index) {
+            inline float& operator[] (const size_t index) {
                 assert(index >= 0 && index < 3);
                 if (index == 0) return x;
                 if (index == 1) return y;
                 return z;
             }
             
-            inline const float& operator[] (const unsigned int index) const {
+            inline const float& operator[] (const size_t index) const {
                 assert(index >= 0 && index < 3);
                 if (index == 0) return x;
                 if (index == 1) return y;
